split the function in many files
