+++ src/message_player_3.c
@@ -1,0 +1,6 @@
+#include "serveur.h"
+
+void	message_player_mort(t_player *p)
+{
+	add_msg_to_player(p, "mort", 4, 1);
+}
